Both labs are complete!
For gridprint, I built on Eric's idea (from class last night) of using strings and for loops, and put them into variables which respond to parameters so you can make the grid as small or big as you'd like.

For fizzbuzz, I used a loop which iterates through each integer in range 0-100, feeds the integer into a string-type variable, and adds the appropriate string into the variable if the modulo checks out.