Sample Scraper
This goes to the BLS page for covered employment data for Washington and reads each line. It selects the lines that contain data for "total nonfarm covered employment", reformats them, and prints them for the user.

You can also use the xlsxwriter module and add a few lines here to write the results in an excel fle. I will post a version that does this later in the week.